fix: resolve repository layer issues
- Fix PaymentRepository: findByCustomerName → findByCustomer_UserName
- Fix PurchaseHistoryRepository: findByCustomerId → findByCustomer_Id

+++ src/main/java/com/example/DifferentWorlds/Repository/AuthorRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 @Repository
-public interface AuthorRepository extends JpaRepository<AuthorEntity,Long> {
+public interface AuthorRepository extends JpaRepository<AuthorEntity,Integer> {
      Optional<AuthorEntity> findByUserName(String UserName);
 }
