+++ src/main/java/com/example/DifferentWorlds/Repository/CustomerRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface CustomerRepository extends JpaRepository<CustomerEntity,Long>{
+public interface CustomerRepository extends JpaRepository<CustomerEntity,Integer>{
      Optional<CustomerEntity> findByUserName(String UserName);
      Optional<CustomerEntity> findByEmail(String email);
 }
